Refactor UI components into shared library
