+++ src/ui/components/index.ts
@@ -6,3 +6,21 @@ export type {
   EventLogFilters,
   EventLogFilterOptions,
 } from "./EventLog";
+export { LaneDiffOverlay } from "./LaneDiffOverlay";
+export type {
+  LaneDiffOverlayProps,
+  LaneDiffOverlayContext,
+  LaneDiffResult,
+  LaneDiffIssue,
+  LaneDiffLagSample,
+  LaneDiffSchemaStatus,
+} from "./LaneDiffOverlay";
+export { MetricsDashboard } from "./MetricsDashboard";
+export type {
+  MetricsDashboardProps,
+  MetricsDashboardLane,
+} from "./MetricsDashboard";
+export { MetricsStrip } from "./MetricsStrip";
+export type { MetricsStripProps } from "./MetricsStrip";
+export { SchemaWalkthrough } from "./SchemaWalkthrough";
+export type { SchemaWalkthroughProps } from "./SchemaWalkthrough";
